# api/openapi.yaml
openapi: 3.0.3
info:
  title: AskAtlas API
  version: 1.0.0
servers:
  - url: /api
security:
  - BearerAuth: []
paths:
...
  /courses:
    get:
      operationId: ListCourses
      summary: List and search courses
      parameters:
        - name: school_id
          in: query
          description: Filter by school UUID
          schema:
            type: string
            format: uuid
        - name: department
          in: query
          description: Exact department code match (case-insensitive, e.g. "CPTS")
          schema:
            type: string
            maxLength: 20
        - name: q
          in: query
          description: Search term, matches against title (trigram), department, or number (case-insensitive)
          schema:
            type: string
            maxLength: 200
        - name: sort_by
          in: query
          description: Database column to sort the results by
          schema:
            type: string
            enum: [department, number, title, created_at]
            default: department
        - name: sort_dir
          in: query
          description: Sorting direction (ascending or descending)
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: page_limit
          in: query
          description: Maximum number of courses to return per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: cursor
          in: query
          description: Opaque pagination cursor token obtained from the previous response
          schema:
            type: string
      responses:
        '200':
          description: A paginated list of courses with embedded school summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCoursesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /courses/{course_id}:
    get:
      operationId: GetCourse
      summary: Get a course detail with embedded sections
      parameters:
        - name: course_id
          in: path
          description: The unique UUID of the course
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The requested course with embedded school summary and sections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /courses/{course_id}/sections:
    get:
      operationId: ListCourseSections
      summary: List sections for a course
      description: |
        Returns every section attached to the given course, with a live
        `member_count` and an optional exact-match `term` filter. Used
        by the course detail page when the caller needs the dedicated
        sections payload (with `course_id` + `created_at`) rather than
        the slimmer inline sections embedded in `GET /courses/{id}`.

        Sorted by `term DESC, section_code ASC` (most-recent term first;
        within a term, section codes ascend). No pagination -- a course
        typically has fewer than 10 sections.

        404 dispatch: a missing course (no row matches `course_id`) is
        a single 404 with "Course not found"; an existing course with
        zero matching sections (filtered out by `term` or just empty)
        returns 200 with `sections: []`. The two are intentionally
        distinguishable so the frontend can show a "no sections in this
        term" empty state vs a generic not-found page.

        Term filter:
          * Exact match (no ILIKE, no trigram). Term values are
            structured strings like "Spring 2026" so case-insensitive
            matching would only mask typos, not real intent.
          * Empty string is treated as "no filter" (defensive: a
            client clearing the input shouldn't send an empty
            string to the server, but if it does the server treats
            it as if no filter was supplied).
      parameters:
        - name: course_id
          in: path
          description: The unique UUID of the parent course
          required: true
          schema:
            type: string
            format: uuid
        - name: term
          in: query
          description: Exact-match filter on the section's term (e.g. "Spring 2026")
          required: false
          schema:
            type: string
            maxLength: 30
      responses:
        '200':
          description: All sections for the course (optionally filtered by term).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCourseSectionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /courses/{course_id}/sections/{section_id}/members:
    get:
      operationId: ListSectionMembers
      summary: List the members of a course section
      description: |
        Returns the section roster with limited per-user info: user_id,
        first_name, last_name, role, joined_at. Email and clerk_id are
        intentionally NOT exposed -- this endpoint is reachable by any
        authenticated user (course pages are public within the app), so
        the payload is the privacy-floor for member identity.

        Sorted by `joined_at ASC` with a `(joined_at, user_id)` keyset
        cursor as the tiebreaker for stable pagination across pages.
      parameters:
        - name: course_id
          in: path
          description: The unique UUID of the parent course
          required: true
          schema:
            type: string
            format: uuid
        - name: section_id
          in: path
          description: The unique UUID of the section
          required: true
          schema:
            type: string
            format: uuid
        - name: role
          in: query
          description: Filter to members with a specific role
          required: false
          schema:
            type: string
            enum:
              - student
              - ta
              - instructor
        - name: limit
          in: query
          description: Maximum number of members to return per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: cursor
          in: query
          description: Opaque keyset cursor token from the previous page
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A paginated list of section members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSectionMembersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: JoinSection
      summary: Join a section as the authenticated user
      description: |
        Adds the authenticated user as a `student` member of the given section.
        The role is hardcoded to `student` regardless of any fields supplied in
        the request body. `instructor` and `ta` roles are assigned only through
        the seeding pipeline.
      parameters:
        - name: course_id
          in: path
          description: The unique UUID of the parent course
          required: true
          schema:
            type: string
            format: uuid
        - name: section_id
          in: path
          description: The unique UUID of the section to join
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Membership created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseMemberResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me/courses:
    get:
      operationId: ListMyEnrollments
      summary: List the authenticated user's section enrollments
      description: |
        Returns every section the authenticated user is enrolled in, with
        compact embedded course + school summaries. Sorted by `term DESC,
        department ASC, number ASC`. Not paginated -- a user is typically
        in 4-8 courses, and even outliers fit comfortably in one response.
      parameters:
        - name: term
          in: query
          description: Exact-match term filter (e.g. "Spring 2026")
          required: false
          schema:
            type: string
            maxLength: 30
        - name: role
          in: query
          description: Filter to a specific course-member role
          required: false
          schema:
            type: string
            enum:
              - student
              - ta
              - instructor
      responses:
        '200':
          description: All enrollments matching the supplied filters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMyEnrollmentsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /courses/{course_id}/sections/{section_id}/members/me:
    get:
      operationId: CheckMembership
      summary: Check the authenticated user's membership in a section
      description: |
        Powers the per-section Join/Leave button on the course detail
        page. Always returns 200 -- non-membership is `enrolled: false`
        with null role/joined_at, NOT 404, so the frontend can
        distinguish "not enrolled" from "section does not exist".
      parameters:
        - name: course_id
          in: path
          description: The unique UUID of the parent course
          required: true
          schema:
            type: string
            format: uuid
        - name: section_id
          in: path
          description: The unique UUID of the section to check
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Membership status for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipCheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: LeaveSection
      summary: Leave a section as the authenticated user
      description: |
        Removes the authenticated user's membership from the given section.
        The membership row is hard-deleted. Any role (`student`, `ta`,
        `instructor`) can leave.
      parameters:
        - name: course_id
          in: path
          description: The unique UUID of the parent course
          required: true
          schema:
            type: string
            format: uuid
        - name: section_id
          in: path
          description: The unique UUID of the section to leave
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Membership removed; no response body
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ListCourseSectionsResponse:
      type: object
      description: |
        Wrapper for GET /courses/{course_id}/sections. Single field
        rather than a bare array so future fields (a filter echo,
        cursor for pagination, aggregate counts) can be added
        backwards-compatibly. Always emits `sections: []` (never
        omitted or null) when the course has no matching sections.
      properties:
        sections:
          type: array
          items:
            $ref: '#/components/schemas/SectionResponse'
      required:
        - sections
    CourseDetailResponse:
      type: object
      description: A course with its school summary and the full list of sections
      properties:
        id:
          type: string
          format: uuid
        school:
          $ref: '#/components/schemas/SchoolSummary'
        department:
          type: string
        number:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        sections:
          type: array
          items:
            $ref: '#/components/schemas/SectionSummary'
      required:
        - id
        - school
        - department
        - number
        - title
        - created_at
        - sections
    ListCoursesResponse:
      type: object
      description: A paginated collection of courses
      properties:
        courses:
          type: array
          items:
            $ref: '#/components/schemas/CourseResponse'
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
      required:
        - courses
        - has_more
    CourseMemberResponse:
      type: object
      description: A user's membership in a course section
      properties:
        user_id:
          type: string
          format: uuid
        section_id:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - student
            - ta
            - instructor
        joined_at:
          type: string
          format: date-time
      required:
        - user_id
        - section_id
        - role
        - joined_at
...
    ListSectionMembersResponse:
      type: object
      description: A paginated collection of section members
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/SectionMemberResponse'
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
      required:
        - members
        - has_more
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'
    Conflict:
      description: Conflict with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppError'